Fix mismatching between TestClient receive logic and GBServer send logic.

+++ GBClient/include/GBClient.h
@@ -28,5 +28,5 @@ public: template<typename STRING> void InscribeStringToBuffer(const STRING& str,
 private: void ClientConsumeGatheredMessage(GBBuffer& buffer, uint32_t receivedMessageSize);
 private: void ConsumeGatheredMessage(ExpandedOverlapped* pEol, char* message, const uint32_t messageSize, int& readOffSet) override {}
 private: void* GetSessionInformationArray() override { return nullptr; }
-private: void GatheringMessage(char* message, uint32_t leftBytesToRecieve);
+private: void GatheringMessage(char* message, uint8_t leftBytesToRecieve);
 };
